#557 add benchmark for JsonUtil perf test
