+++ coffee-tool/pom.xml
@@ -106,5 +106,13 @@
 			<groupId>org.eclipse</groupId>
 			<artifactId>yasson</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.openjdk.jmh</groupId>
+			<artifactId>jmh-core</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.openjdk.jmh</groupId>
+			<artifactId>jmh-generator-annprocess</artifactId>
+		</dependency>
 	</dependencies>
 </project>
