+++ coffee-tool/src/test/java/hu/icellmobilsoft/coffee/tool/utils/json/JsonUtilTest.java
@@ -23,17 +23,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import java.time.LocalDate;
-import java.time.OffsetDateTime;
-import java.time.OffsetTime;
 import java.time.YearMonth;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
-import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
-import java.util.Objects;
 import java.util.TimeZone;
 
 import javax.xml.datatype.DatatypeConfigurationException;
@@ -50,9 +45,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
+
 import hu.icellmobilsoft.coffee.se.api.exception.BaseException;
 import hu.icellmobilsoft.coffee.se.api.exception.JsonConversionException;
-
 
 /**
  * @author mark.petrenyi
@@ -142,24 +137,24 @@ class JsonUtilTest {
             @Test
             @DisplayName("Testing toObject() with invalid JSON object YearMonth")
             void invalidJsonObjectYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 TestObject actual = JsonUtil.toObjectOpt(INVALID_JSON_OBJECT_YEAR_MONTH, TestObject.class).orElse(null);
 
-                //then
+                // then
                 assertNull(actual);
             }
 
             @Test
             @DisplayName("Testing toObject() with invalid JSON array YearMonth")
             void invalidJsonArrayYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 TestObject actual = JsonUtil.toObjectOpt(INVALID_JSON_ARRAY_YEAR_MONTH, TestObject.class).orElse(null);
 
-                //then
+                // then
                 assertNull(actual);
             }
 
@@ -222,24 +217,24 @@ class JsonUtilTest {
             @Test
             @DisplayName("Testing toObjectEx() with invalid JSON object YearMonth")
             void invalidJsonObjectYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 Executable operation = () -> JsonUtil.toObjectEx(INVALID_JSON_OBJECT_YEAR_MONTH, TestObject.class);
 
-                //then
+                // then
                 assertThrows(BaseException.class, operation);
             }
 
             @Test
             @DisplayName("Testing toObjectEx() with invalid JSON array YearMonth")
             void invalidJsonArrayYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 Executable operation = () -> JsonUtil.toObjectEx(INVALID_JSON_ARRAY_YEAR_MONTH, TestObject.class);
 
-                //then
+                // then
                 assertThrows(BaseException.class, operation);
             }
         }
@@ -278,24 +273,24 @@ class JsonUtilTest {
             @Test
             @DisplayName("Testing toObjectGson with invalid JSON object YearMonth")
             void invalidJsonObjectYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 Executable operation = () -> JsonUtil.toObjectGson(INVALID_JSON_OBJECT_YEAR_MONTH, TestObject.class);
 
-                //then
+                // then
                 assertThrows(Exception.class, operation);
             }
 
             @Test
             @DisplayName("Testing toObjectGson with invalid JSON array YearMonth")
             void invalidJsonArrayYearMonth() {
-                //given
+                // given
 
-                //when
+                // when
                 Executable operation = () -> JsonUtil.toObjectGson(INVALID_JSON_ARRAY_YEAR_MONTH, TestObject.class);
 
-                //then
+                // then
                 assertThrows(Exception.class, operation);
             }
         }
@@ -430,135 +425,6 @@ class JsonUtilTest {
                 assertThrows(Exception.class, operation);
             }
         }
-    }
-
-    /**
-     * Helper class for testing
-     */
-    public static class TestObject {
-        XMLGregorianCalendar xmlGregorianCalendar;
-        Date date;
-        byte[] bytes;
-        String string;
-        Class<?> clazz;
-        OffsetDateTime offsetDateTime;
-        OffsetTime offsetTime;
-        LocalDate localDate;
-        Duration duration;
-        YearMonth yearMonth;
-
-        public XMLGregorianCalendar getXmlGregorianCalendar() {
-            return xmlGregorianCalendar;
-        }
-
-        public void setXmlGregorianCalendar(XMLGregorianCalendar xmlGregorianCalendar) {
-            this.xmlGregorianCalendar = xmlGregorianCalendar;
-        }
-
-        public Date getDate() {
-            return date;
-        }
-
-        public void setDate(Date date) {
-            this.date = date;
-        }
-
-        public byte[] getBytes() {
-            return bytes;
-        }
-
-        public void setBytes(byte[] bytes) {
-            this.bytes = bytes;
-        }
-
-        public String getString() {
-            return string;
-        }
-
-        public void setString(String string) {
-            this.string = string;
-        }
-
-        public Class<?> getClazz() {
-            return clazz;
-        }
-
-        public void setClazz(Class<?> clazz) {
-            this.clazz = clazz;
-        }
-
-        public OffsetDateTime getOffsetDateTime() {
-            return offsetDateTime;
-        }
-
-        public void setOffsetDateTime(OffsetDateTime offsetDateTime) {
-            this.offsetDateTime = offsetDateTime;
-        }
-
-        public OffsetTime getOffsetTime() {
-            return offsetTime;
-        }
-
-        public void setOffsetTime(OffsetTime offsetTime) {
-            this.offsetTime = offsetTime;
-        }
-
-        public LocalDate getLocalDate() {
-            return localDate;
-        }
-
-        public void setLocalDate(LocalDate localDate) {
-            this.localDate = localDate;
-        }
-
-        public Duration getDuration() {
-            return duration;
-        }
-
-        public void setDuration(Duration duration) {
-            this.duration = duration;
-        }
-
-        public YearMonth getYearMonth() {
-            return yearMonth;
-        }
-
-        public void setYearMonth(YearMonth yearMonth) {
-            this.yearMonth = yearMonth;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (!(o instanceof TestObject)) {
-                return false;
-            }
-            TestObject other = (TestObject) o;
-
-            return Objects.equals(xmlGregorianCalendar, other.xmlGregorianCalendar) //
-                    && Objects.equals(date, other.date)//
-                    && Objects.deepEquals(bytes, other.bytes)//
-                    && Objects.equals(string, other.string)//
-                    && Objects.equals(clazz, other.clazz) && Objects.equals(offsetDateTime, other.offsetDateTime)
-                    && Objects.equals(offsetTime, other.offsetTime) && Objects.equals(localDate, other.localDate)
-                    && Objects.equals(duration, other.duration)
-                    && Objects.equals(yearMonth, other.yearMonth);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(xmlGregorianCalendar, date, bytes, string, clazz, offsetDateTime, offsetTime, localDate, duration, yearMonth);
-        }
-
-        @Override
-        public String toString() {
-            return "TestObject{" + "xmlGregorianCalendar=" + xmlGregorianCalendar + ", date=" + date + ", bytes=" + Arrays.toString(bytes)
-                    + ", string='" + string + '\'' + ", clazz=" + clazz + "offsetDateTime=" + offsetDateTime + ", offsetTime=" + offsetTime
-                    + ", localDate=" + localDate + ", duration=" + duration + ", yearMonth=" + yearMonth + '}';
-        }
-
     }
 
 }
